Rename Nodes package to avoid import collision
