+++ node_ext/military_nodes/index.go
